MGS-9035 gpu: Integrate Mali DDK r56p0
- Allowed flags with kbase iotcl calls
- Remove legacy IOCTLs KBASE_IOCTL_SET_LIMITED_CORE_COUNT and KBASE_IOCTL_DISJOINT_QUERY

+++ drivers/gpu/arm/midgard/mali_kbase_dvfs_debugfs.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0 WITH Linux-syscall-note
 /*
  *
- * (C) COPYRIGHT 2020-2024 ARM Limited. All rights reserved.
+ * (C) COPYRIGHT 2020-2025 ARM Limited. All rights reserved.
  *
  * This program is free software and is provided to you under the terms of the
  * GNU General Public License version 2 as published by the Free Software
@@ -39,9 +39,9 @@ static int kbasep_dvfs_utilization_debugfs_show(struct seq_file *file, void *dat
 
 	CSTD_UNUSED(data);
 	seq_printf(file, "busy_time: %llu idle_time: %llu protm_time: %llu\n",
-		   kbdev->pm.backend.metrics.values.time_busy,
-		   kbdev->pm.backend.metrics.values.time_idle,
-		   kbdev->pm.backend.metrics.values.time_in_protm);
+		   (unsigned long long)kbdev->pm.backend.metrics.values.time_busy,
+		   (unsigned long long)kbdev->pm.backend.metrics.values.time_idle,
+		   (unsigned long long)kbdev->pm.backend.metrics.values.time_in_protm);
 
 	seq_printf(file, "shader_frag_busy_time: %llu shader_compute_busy_time: %llu tiler_busy_time: %llu\n",
 		   kbdev->pm.backend.metrics.values.shader_frag_time_busy,
